requirements.txt
PyPDF2>=3.0.1
python-docx>=1.1.0
Pillow>=10.0.0
openpyxl>=3.1.0
pandas>=2.0.0
pdfplumber>=0.11.0
pytesseract>=0.3.0
pdf2image>=1.17.0
pyinstaller>=6.0.0==3.0.1
python-docx==1.1.0
Pillow>=10.0.0
openpyxl>=3.1.0
pandas>=2.0.0
pyinstaller>=6.0.0